backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
boto3==1.29.7
psycopg2-binary>=2.9.9
sqlalchemy==2.0.23
geoalchemy2==0.14.2
shapely>=2.0.2
# Geospatial packages requiring system dependencies - install after PROJ/GDAL setup (see GDAL_INSTALL_WINDOWS.md)
# geopandas==0.14.1
# rasterio==1.3.9
# fiona==1.9.5
# pyproj==3.6.1
numpy>=1.26.0,<2.0.0
pandas>=2.2.0,<2.3.0
scipy>=1.12.0,<1.13.0
scikit-learn>=1.4.0,<1.5.0
openai==1.3.7
httpx==0.25.2
pyyaml==6.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
reportlab==4.0.7
pillow==10.1.0

